% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/indicator.R, R/indicator_area.R
\name{authority}
\alias{authority}
\alias{authority.indicator}
\alias{authority.indicator_area}
\title{Get authority of object.}
\usage{
authority(x)

\method{authority}{indicator}(x)

\method{authority}{indicator_area}(x)
}
\arguments{
\item{x}{Appropriate object e.g. indicator_area/indicator object.}
}
\description{
Get authority of object.
}
