% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/indicator.R
\name{category}
\alias{category}
\alias{category.indicator}
\title{Get category of object.}
\usage{
category(x)

\method{category}{indicator}(x)
}
\arguments{
\item{x}{Appropriate object e.g. indicator object.}
}
\description{
Get category of object.
}
