% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/indicator.R, R/indicator_area.R
\name{dataset}
\alias{dataset}
\alias{dataset.indicator}
\alias{dataset.indicator_area}
\title{Get data set of object.}
\usage{
dataset(x)

\method{dataset}{indicator}(x)

\method{dataset}{indicator_area}(x)
}
\arguments{
\item{x}{Appropriate object e.g. indicator_area/indicator object.}
}
\description{
Get data set of object.
}
