% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/indicator_area.R
\name{datatable_server}
\alias{datatable_server}
\alias{datatable_server.indicator_area}
\title{Get server side code for datatable for object.}
\usage{
datatable_server(x, ...)

\method{datatable_server}{indicator_area}(x, ...)
}
\arguments{
\item{x}{Appropriate object e.g. indicator_area object.}

\item{...}{Passed to methods.}
}
\description{
Get server side code for datatable for object.
}
