% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/indicator_area.R
\name{datatable_ui}
\alias{datatable_ui}
\alias{datatable_ui.indicator_area}
\title{Get user interface containing datatable for object.}
\usage{
datatable_ui(x, ...)

\method{datatable_ui}{indicator_area}(x, ...)
}
\arguments{
\item{x}{Appropriate object e.g. indicator_area object.}

\item{...}{Passed to methods.}
}
\description{
Get user interface containing datatable for object.
}
