% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_indicator_data}
\alias{example_indicator_data}
\title{Example indicator data}
\description{
Data.frame object containing indicator data for 2 indicators across
2 different local authorities
}
\keyword{data}
