% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/indicator.R
\name{id}
\alias{id}
\alias{id.default}
\alias{id.indicator}
\title{Get id of object.}
\usage{
id(x)

\method{id}{default}(x)

\method{id}{indicator}(x)
}
\arguments{
\item{x}{Appropriate object e.g. indicator_area/indicator object.}
}
\description{
Get id of object.
}
