% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{indicator}
\alias{indicator}
\title{Create indicator object}
\usage{
indicator(x, ...)
}
\arguments{
\item{x}{data.frame object indicator data for a single local authority.}

\item{...}{Passed to new_tibble.}
}
\description{
Indicator objects contain a data.frame of data for a single indicator in
a single local authority
}
