% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_area.R
\name{indicator_area}
\alias{indicator_area}
\title{Create indicator area object}
\usage{
indicator_area(x, id, ...)
}
\arguments{
\item{x}{list of indicator objects.}

\item{id}{Character id for object.}

\item{...}{Passed to structure}
}
\description{
Indicator area objects contain a list of indicator objects for single
indicator serivce area.
}
