% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_area.R
\name{is_indicator_area}
\alias{is_indicator_area}
\title{Is object indicator area object?}
\usage{
is_indicator_area(x)
}
\arguments{
\item{x}{object to test.}
}
\description{
Is object indicator area object?
}
