% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/indicator_area.R
\name{menu_item}
\alias{menu_item}
\alias{menu_item.indicator_area}
\title{Get menu item user interface for object.}
\usage{
menu_item(x)

\method{menu_item}{indicator_area}(x, ...)
}
\arguments{
\item{x}{Appropriate object e.g. indicator_area object.}

\item{...}{Passed to methods.}
}
\description{
Get menu item user interface for object.
}
