% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/indicator.R
\name{plot_data}
\alias{plot_data}
\alias{plot_data.indicator}
\title{Get plot data for object.}
\usage{
plot_data(x, ...)

\method{plot_data}{indicator}(x, type, ...)
}
\arguments{
\item{x}{Appropriate object e.g. indicator object.}

\item{...}{Passed to methods.}

\item{type}{character type of plot to get data for.}
}
\description{
Get plot data for object.
}
