% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/indicator.R
\name{plot_types}
\alias{plot_types}
\alias{plot_types.indicator}
\title{Get available plot types for object.}
\usage{
plot_types(x)

\method{plot_types}{indicator}(x)
}
\arguments{
\item{x}{Appropriate object e.g. indicator object.}
}
\description{
Get available plot types for object.
}
