% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/indicator.R
\name{title}
\alias{title}
\alias{title.indicator}
\title{Get title of object.}
\usage{
title(x)

\method{title}{indicator}(x)
}
\arguments{
\item{x}{Appropriate object e.g. indicator object.}
}
\description{
Get title of object.
}
