% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/indicator.R, R/indicator_area.R
\name{validate}
\alias{validate}
\alias{validate.indicator}
\alias{validate.indicator_area}
\title{Validate object has correct stucture.}
\usage{
validate(x, ...)

\method{validate}{indicator}(x, ...)

\method{validate}{indicator_area}(x, ...)
}
\arguments{
\item{x}{Object to validate e.g. indicator_area/indicator object.}

\item{...}{Passed to methods}
}
\description{
Validate object has correct stucture.
}
