% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/indicator.R, R/indicator_area.R
\name{visual_server}
\alias{visual_server}
\alias{visual_server.indicator}
\alias{visual_server.indicator_area}
\title{Get server side code for visualisation for object.}
\usage{
visual_server(x, ...)

\method{visual_server}{indicator}(x, ...)

\method{visual_server}{indicator_area}(x, ...)
}
\arguments{
\item{x}{Appropriate object e.g. indicator_area/indicator object.}

\item{...}{Passed to methods.}
}
\description{
Get server side code for visualisation for object.
}
