% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/indicator.R, R/indicator_area.R
\name{visual_ui}
\alias{visual_ui}
\alias{visual_ui.indicator}
\alias{visual_ui.indicator_area}
\title{Get user interface containing visualisation for object.}
\usage{
visual_ui(x, ...)

\method{visual_ui}{indicator}(x, ns, ...)

\method{visual_ui}{indicator_area}(x, ...)
}
\arguments{
\item{x}{Appropriate object e.g. indicator_area/indicator object.}

\item{...}{Passed to methods.}

\item{ns}{Namespace function to generate UI with.}
}
\description{
Get user interface containing visualisation for object.
}
